move the "paste" menu item to the edit fragment and call coerceToText on the clipboard content...
