+++ sholi/src/main/res/menu/checking.xml
@@ -26,12 +26,6 @@
         android:orderInCategory="1"
         android:showAsAction="ifRoom" />
 
-    <item android:id="@+id/action_paste"
-        android:icon="@android:drawable/ic_menu_agenda"
-        android:title="@string/action_checking_paste"
-        android:orderInCategory="1"
-        android:showAsAction="ifRoom" />
-
     <item android:id="@+id/action_menu"
         android:icon="@android:drawable/ic_menu_manage"
         android:title="@string/action_checking_menu"
