+++ sholi/src/main/res/menu/edit.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".MainActivity">
+
+    <item
+        android:id="@+id/action_paste"
+        android:icon="@android:drawable/ic_menu_agenda"
+        android:orderInCategory="1"
+        android:showAsAction="ifRoom"
+        android:title="@string/action_edit_paste" />
+
+</menu>
